+++ samplesheets/tests/test_tasks.py
@@ -9,12 +9,25 @@ from unittest import skipIf
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
 from projectroles.models import SODAR_CONSTANTS
-from projectroles.tests.test_views_api import SODARAPIViewTestMixin
+from projectroles.plugins import get_backend_api
 from projectroles.tests.test_views_taskflow import TestTaskflowBase
 
+# Sodarcache dependency
+from sodarcache.models import JSONCacheItem
+
+# Appalerts dependency
+from appalerts.models import AppAlert
+
+# Timeline dependency
+from timeline.models import ProjectEvent
+
 from samplesheets.models import ISATab
-from samplesheets.tasks import sheet_sync_task
+from samplesheets.tasks import update_project_cache_task, sheet_sync_task
 from samplesheets.tests.test_io import SampleSheetIOMixin, SHEET_DIR
+from samplesheets.tests.test_views_taskflow import (
+    SampleSheetTaskflowMixin,
+    TestSheetSyncBase,
+)
 
 
 app_settings = AppSettingAPI()
@@ -35,90 +48,114 @@ SUBMIT_STATUS_PENDING_TASKFLOW = SODAR_CONSTANTS[
 ]
 
 # Local constants
+APP_NAME = 'samplesheets'
 SHEET_PATH = SHEET_DIR + 'i_small.zip'
+CACHE_ALERT_MESSAGE = 'Testing'
 TASKFLOW_ENABLED = (
     True if 'taskflow' in settings.ENABLED_BACKEND_PLUGINS else False
 )
 TASKFLOW_SKIP_MSG = 'Taskflow not enabled in settings'
-APP_NAME = 'samplesheets'
-ZONE_TITLE = '20190703_172456'
-ZONE_SUFFIX = 'Test Zone'
-ZONE_DESC = 'description'
-TEST_OBJ_NAME = 'test1.txt'
-ASYNC_WAIT_SECONDS = 5
-ASYNC_RETRY_COUNT = 3
+IRODS_BACKEND_ENABLED = (
+    True if 'omics_irods' in settings.ENABLED_BACKEND_PLUGINS else False
+)
+IRODS_BACKEND_SKIP_MSG = 'iRODS backend not enabled in settings'
 
 
-class TestSheetSyncBase(
-    SODARAPIViewTestMixin,
-    SampleSheetIOMixin,
-    TestTaskflowBase,
+@skipIf(not IRODS_BACKEND_ENABLED, IRODS_BACKEND_SKIP_MSG)
+class TestUpdateProjectCacheTask(
+    SampleSheetIOMixin, SampleSheetTaskflowMixin, TestTaskflowBase
 ):
+    """Tests for project cache update task"""
+
     def setUp(self):
         super().setUp()
-
-        # Make owner user
-        self.user_owner_source = self.make_user('owner_source')
-        self.user_owner_target = self.make_user('owner_target')
-
-        # Create Projects
-        self.project_source, self.owner_as_source = self._make_project_taskflow(
-            title='TestProjectSource',
+        self.project, self.owner_as = self._make_project_taskflow(
+            title='TestProject',
             type=PROJECT_TYPE_PROJECT,
             parent=self.category,
-            owner=self.user_owner_source,
+            owner=self.user,
             description='description',
         )
-        self.project_target, self.owner_as_target = self._make_project_taskflow(
-            title='TestProjectTarget',
-            type=PROJECT_TYPE_PROJECT,
-            parent=self.category,
-            owner=self.user_owner_target,
-            description='description',
+        self.investigation = self._import_isa_from_file(
+            SHEET_PATH, self.project
         )
+        self.study = self.investigation.studies.first()
+        self.assay = self.study.assays.first()
+        self.app_alerts = get_backend_api('appalerts_backend')
+        self._make_irods_colls(self.investigation)
 
-        # Import investigation
-        self.inv_source = self._import_isa_from_file(
-            SHEET_PATH, self.project_source
-        )
-
-        self.p_id_source = 'p{}'.format(self.project_source.pk)
-        self.p_id_target = 'p{}'.format(self.project_target.pk)
-
-        # Allow sample sheet editing in project
-        app_settings.set_app_setting(
-            APP_NAME, 'sheet_sync_enable', True, project=self.project_target
-        )
-        app_settings.set_app_setting(
-            APP_NAME,
-            'sheet_sync_url',
-            self.live_server_url
-            + reverse(
-                'samplesheets:api_export_json',
-                kwargs={'project': str(self.project_source.sodar_uuid)},
-            ),
-            project=self.project_target,
-        )
-        app_settings.set_app_setting(
-            APP_NAME,
-            'sheet_sync_token',
-            self.get_token(self.user_owner_source),
-            project=self.project_target,
-        )
-
-        # Check if source is set up correctly
-        self.assertEqual(self.project_source.investigations.count(), 1)
+    def test_update_cache(self):
+        """Test cache update"""
         self.assertEqual(
-            self.project_source.investigations.first().studies.count(), 1
+            JSONCacheItem.objects.filter(project=self.project).count(), 0
         )
+        self.assertEqual(AppAlert.objects.all().count(), 1)
+        self.assertEqual(ProjectEvent.objects.all().count(), 1)
+
+        update_project_cache_task(
+            self.project.sodar_uuid,
+            self.user.sodar_uuid,
+            add_alert=True,
+            alert_msg=CACHE_ALERT_MESSAGE,
+        )
+
         self.assertEqual(
-            self.project_source.investigations.first()
-            .studies.first()
-            .assays.count(),
-            1,
+            JSONCacheItem.objects.filter(project=self.project).count(), 1
         )
-        self.assertEqual(self.project_target.investigations.count(), 0)
-        self.assertEqual(ISATab.objects.count(), 1)
+        cache_item = JSONCacheItem.objects.first()
+        self.assertEqual(
+            cache_item.name,
+            'irods/shortcuts/assay/{}'.format(self.assay.sodar_uuid),
+        )
+        expected_data = {
+            'shortcuts': {
+                'misc_files': False,
+                'track_hubs': [],
+                'results_reports': False,
+            }
+        }
+        self.assertEqual(cache_item.data, expected_data)
+        self.assertEqual(AppAlert.objects.all().count(), 2)
+        alert = AppAlert.objects.order_by('-pk').first()
+        self.assertTrue(alert.message.endswith(CACHE_ALERT_MESSAGE))
+        self.assertEqual(ProjectEvent.objects.all().count(), 2)
+
+    def test_update_cache_no_alert(self):
+        """Test cache update with app alert disabled"""
+        self.assertEqual(
+            JSONCacheItem.objects.filter(project=self.project).count(), 0
+        )
+        self.assertEqual(AppAlert.objects.all().count(), 1)
+        self.assertEqual(ProjectEvent.objects.all().count(), 1)
+
+        update_project_cache_task(
+            self.project.sodar_uuid, self.user.sodar_uuid, add_alert=False
+        )
+
+        self.assertEqual(
+            JSONCacheItem.objects.filter(project=self.project).count(), 1
+        )
+        self.assertEqual(AppAlert.objects.all().count(), 1)
+        self.assertEqual(ProjectEvent.objects.all().count(), 2)
+
+    def test_update_cache_no_user(self):
+        """Test cache update with no user"""
+        self.assertEqual(
+            JSONCacheItem.objects.filter(project=self.project).count(), 0
+        )
+        self.assertEqual(AppAlert.objects.all().count(), 1)
+        self.assertEqual(ProjectEvent.objects.all().count(), 1)
+
+        update_project_cache_task(self.project.sodar_uuid, None, add_alert=True)
+
+        self.assertEqual(
+            JSONCacheItem.objects.filter(project=self.project).count(), 1
+        )
+        self.assertEqual(AppAlert.objects.all().count(), 1)
+        self.assertEqual(ProjectEvent.objects.all().count(), 2)
+
+
+# NOTE: TestSheetSyncBase moved to test_views_taskflow due to circular import
 
 
 @skipIf(not TASKFLOW_ENABLED, TASKFLOW_SKIP_MSG)