+++ samplesheets/views.py
@@ -426,6 +426,8 @@ class SampleSheetImportMixin:
             update_project_cache_task.delay(
                 project_uuid=str(project.sodar_uuid),
                 user_uuid=str(self.request.user.sodar_uuid),
+                add_alert=ui_mode,
+                alert_msg='Sample sheet {}d'.format(action),
             )
             if ui_mode:
                 success_msg += ', initiated iRODS cache update'
@@ -671,6 +673,8 @@ class IrodsCollsCreateViewMixin:
             update_project_cache_task.delay(
                 project_uuid=str(project.sodar_uuid),
                 user_uuid=str(self.request.user.sodar_uuid),
+                add_alert=True,
+                alert_msg='iRODS collection {}'.format(action),
             )
 
         # If public guest access and anonymous allowed, add ticket access
@@ -923,11 +927,13 @@ class SheetSync(SampleSheetImportMixin):
             update_project_cache_task(
                 project_uuid=str(project.sodar_uuid),
                 user_uuid=str(user.sodar_uuid),
+                add_alert=True,
+                alert_msg='Remote sample sheets synchronized',
             )
 
         logger.info(
-            f'Sample sheet sync OK for project "{project.title}" '
-            f'({project.sodar_uuid})'
+            'Sample sheet sync OK for project "{}" '
+            '({})'.format(project.title, project.sodar_uuid)
         )
         return True
 
@@ -1451,11 +1457,14 @@ class SampleSheetCacheUpdateView(
             update_project_cache_task.delay(
                 project_uuid=str(project.sodar_uuid),
                 user_uuid=str(request.user.sodar_uuid),
+                add_alert=True,
+                alert_msg='Manual cache update',
             )
             messages.warning(
                 self.request,
-                'Cache updating initiated. This may take some time, refresh '
-                'sheet view after a while to see the results.',
+                'Cache updating initiated. This may take some time, you will '
+                'receive an alert once done. Refresh the sheet view to see '
+                'the results.',
             )
         return redirect(get_sheets_url(project))