+++ CHANGELOG.rst
@@ -17,10 +17,12 @@ Added
     - Siteinfo app in default ``LOGGING_APPS`` value (#1219)
 - **Landingzones**
     - ``busyzones`` management command (#1212)
+    - App alerts for UI sheet cache updates (#1000)
 - **Samplesheets**
     - Simple link support for string cell rendering (#1001)
     - ``generic_raw`` assay plugin (#1128)
     - Overriding assay plugin via assay comment (#1128)
+    - App alerts for UI sheet cache updates (#1000)
 
 Changed
 -------
@@ -35,6 +37,8 @@ Changed
 - **Irodsinfo**
     - Retrieve iRODS config from ``IRODS_ENV_CLIENT`` setting (#1221)
     - Display ``IRODS_ENV_CLIENT`` in siteinfo via ``info_settings``
+- **Samplesheets**
+    - Move ``TestSheetSyncBase`` into ``test_views_taskflow``
 
 Fixed
 -----